Allow to follow paths on subprojects
