+++ scrunch/order.py
@@ -563,46 +563,3 @@ class ProjectDatasetsOrder(Order):
     def _load(self, refresh=False):
         self.datasets = self.catalog.by('id')
         super(ProjectDatasetsOrder, self)._load(refresh=refresh)
-
-
-class NestedGroup(object):
-    pass
-
-
-class NestedProjectsOrder(object):
-    def __init__(self, project_res):
-        self.project_res = project_res
-
-    def create_project(self, name):
-        # This should be a method of the Project class
-        from scrunch.datasets import Project
-        proj_res = self.project_res.create(shoji_entity_wrapper({
-            'name': name
-        }))
-        return Project(proj_res)
-
-    def get(self, path):
-        self.project_res.refresh()  # Always up to date
-        node = self
-        for p_name in Path(path).get_parts():
-            try:
-                node = node.order.get_child(p_name)
-            except KeyError:
-                raise InvalidPathError('Project not found %s' % p_name)
-        return node
-
-    def __getitem__(self, path):
-        return self.get(path)
-
-    def get_child(self, name):
-        from scrunch.datasets import Project
-        by_name = self.project_res.by('name')
-
-        if name in by_name:
-            # Found by name, if it's not a folder, return the variable
-            tup = by_name[name]
-            if tup.type == 'project':
-                return Project(tup.entity)
-            return self.root.dataset[name]
-
-        raise InvalidPathError('Invalid path: %s' % name)
